Allow lua code to attach things to 'self'
Instead of passing `this` directly as the first param, pass a table attached
to the instance with an __index metamethod that tries the lookup in the delegate
table and only tries the clr object as a last resort.

The metamethod is implemented in Lua so we don't have to reimplement anything
goofy like recursing through multiple metatables or trying to figure out the
right method to call on the clr object.